requirements.txt
certifi==2020.12.5
cftime>=1.4.1
netCDF4>=1.5.6
numpy>=1.19.5
pandas>=1.1.5
python-dateutil>=2.8.1
pytz>=2021.1
scipy>=1.5.4
six>=1.15.0
wrapt>=1.12.1
wrf-python>=1.2.0
xarray>=0.16.2
